requirements.txt
Django==2.0.3
djangorestframework==3.7.7
dj-database-url==0.5.0
pandas==0.22.0
psycopg2==2.7.4
sqlalchemy==1.2.5
xlrd==1.1.0
